prompt templates added and code refactoring
