+++ src/aem.commands.ts
@@ -1,7 +1,7 @@
 export enum AEM_COMMANDS {
     CREATE = "create",
     INFO = "info",
-    ENHANCE = "enhance",
+    // ENHANCE = "enhance", //TODO
     COLLECION = "collection",
     DOCS = "docs",
 }
