+++ src/extension.ts
@@ -4,15 +4,13 @@ import {
   createCmdHandler,
 } from "./handlers/block.create";
 import { infoCmdHandler } from "./handlers/block.info";
-import { enhanceBlock as enhanceCmdHandler } from "./handlers/block.enhancer";
-import { defaultHandler } from "./handlers/default";
 import { handleDocsCommand } from "./handlers/block.docs";
 
 import {
   AEM_COMMAND_ID,
   PROCESS_COPILOT_CREATE_CMD,
 } from "./constants";
-import { createFolderAndFiles } from "./utils";
+import { createFolderAndFiles } from "./utils/fileHandler";
 import { fetchBlock } from "./handlers/block.collections";
 
 interface IAemChatResult extends vscode.ChatResult {
@@ -35,14 +33,10 @@ export function activate(context: vscode.ExtensionContext) {
         cmdResult = await infoCmdHandler(request, stream, token);
       } else if (request.command == commands.CREATE) {
         cmdResult = await createCmdHandler(request, stream, token);
-      } else if (request.command == commands.ENHANCE) {
-        cmdResult = await enhanceCmdHandler(request, stream, token);
       } else if (request.command == commands.COLLECION) {
         cmdResult = await fetchBlock(request, stream, token, context);
-      } else if (request.command == commands.DOCS) {
-        cmdResult = await handleDocsCommand(request, stream, token);
       } else {
-        cmdResult = await defaultHandler(request, stream, token);
+        cmdResult = await handleDocsCommand(request, stream, token);
       }
     } catch (err) {
       handleError(err, stream);
