+++ src/interfaces/fileinterface.ts
@@ -1,0 +1,5 @@
+export interface File {
+    name: string;
+    path: string;
+    content?: string;
+}
